requirements.txt
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
loguru==0.7.3
mysql-connector-python==9.2.0
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
pycparser==2.22
pypdfium2==4.30.1
python-dotenv==1.0.1
win32_setctime==1.2.0
